projects/Gaia/requirements.txt
black==26.3.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
durationpy==0.10
idna==3.11
kubernetes==35.0.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
python-dateutil==2.9.0.post0
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
urllib3==2.6.3
websocket-client==1.9.0
